+++ components/elements/AnimatedText.js
@@ -1,11 +1,7 @@
-import React from "react";
-
 export default function AnimatedText() {
   return (
     <div className="text-white hidden md:block sm:text-2xl md:text-3xl lg:text-4xl 2xl:text-5xl font-medium overflow-y-hidden">
-      Create a {/* <TextLoop> */}
-      <span className="text-white ">Near Social</span>
-      Profile
+      Create a <span className="text-white">Near Social</span> Profile
     </div>
   );
 }
